Добавил файл trip-day.js в новой папке controllers и отредактировал его работу в main.js
